[EI-218] [api] refactor: 파일 경로 변경
