+++ api/src/numerical-guidance/application/port/persistence/indicator/load-indicator.port.ts
@@ -1,4 +1,4 @@
-import { IndicatorDto } from '../../../query/indicator/dto/indicator.dto';
+import { IndicatorDto } from '../../../query/indicator/basic/dto/indicator.dto';
 
 export interface LoadIndicatorPort {
   loadIndicator(id: string): Promise<IndicatorDto>;
