+++ api/src/numerical-guidance/application/port/persistence/indicator/load-indicators.port.ts
@@ -1,4 +1,4 @@
-import { IndicatorsDto } from '../../../query/indicator/dto/indicators.dto';
+import { IndicatorsDto } from '../../../query/indicator/basic/dto/indicators.dto';
 
 export interface LoadIndicatorsPort {
   loadIndicators(): Promise<IndicatorsDto>;
